fftech_website_audit_saas/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[argon2]
python-jose
pydantic
apscheduler
requests
beautifulsoup4
reportlab
python-multipart
